Implement call and jmp

+++ simulator/include/types.h
@@ -1,4 +1,4 @@
 #pragma once
 
-using address_t = uint32_t;
+using address_t = uint16_t;
 using byte_t = uint8_t;
